refactor(kolme): extract finality checker module ownership (#1946)
